Tile: add delta
For mmol/L conversion at the edges this probably does the wrong thing (unlike the NightScout web UI) - it's possible for a 1mg/dL delta to lead to no change in the rounded converted value. Though it's somewhat useful to know nightscout saw a small change in its native units, it's unintuitive to most end users if the blood glucose reading is unchanged in mmol/L, but we show a 0.1mmol/L delta.

As a concrete example, if we have the sequence

    192mg/dL, 193mg/dL

this will be converted and rounded to

    10.7mmol/L, 10.7mmol/L

but the +1mg/dL delta will be converted and rounded to +0.1mmol/L.

The only apparent way round this is to derive the delta after conversion...